rename middleware folder and update imports
